# docs/api/openapi.json
{
  "components": {
...
    "securitySchemes": {
      "session": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "TLS, exact origin, client CIDR policy and expiring opaque bearer sessions. No cookies. Disconnected RouterOS/subscription adapters return 501; standalone apply returns 503.",
    "title": "MikroCentauri API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "operationId": "post_auth_login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "password": {
                    "maxLength": 1024,
                    "type": "string"
                  }
                },
                "required": [
                  "password"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Operation completed"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "409": {
            "description": "Stale draft/plan"
          },
          "422": {
            "description": "Validation failed"
          },
          "429": {
            "description": "Login limit"
          },
          "503": {
            "description": "Runtime or persistence unavailable"
          }
        },
        "security": []
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "post_auth_logout",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "description": "Operation completed"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "409": {
            "description": "Stale draft/plan"
          },
          "422": {
            "description": "Validation failed"
          },
          "429": {
            "description": "Login limit"
          },
          "503": {
            "description": "Runtime or persistence unavailable"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/v1/backup": {
      "get": {
        "operationId": "get_backup",
        "responses": {
          "200": {
            "description": "Redacted resource"
          },
          "401": {
            "description": "Authentication required"
          },
          "503": {
            "description": "Not ready or unavailable"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
    "/api/v1/config": {
      "get": {
        "operationId": "get_config",
        "responses": {
          "200": {
            "description": "Redacted resource"
          },
          "401": {
            "description": "Authentication required"
          },
          "503": {
            "description": "Not ready or unavailable"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/v1/config/apply": {
      "post": {
        "operationId": "post_config_apply",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "plan_id": {
                    "maxLength": 64,
                    "minLength": 64,
                    "type": "string"
                  }
                },
                "required": [
                  "plan_id"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Operation completed"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "409": {
            "description": "Stale draft/plan"
          },
          "422": {
            "description": "Validation failed"
          },
          "429": {
            "description": "Login limit"
          },
          "503": {
            "description": "Runtime or persistence unavailable"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/v1/config/draft": {
      "get": {
        "operationId": "get_config_draft",
        "responses": {
          "200": {
            "description": "Redacted resource"
          },
          "401": {
            "description": "Authentication required"
          },
          "503": {
            "description": "Not ready or unavailable"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "post": {
        "operationId": "post_config_draft",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CoreModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Operation completed"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "409": {
            "description": "Stale draft/plan"
          },
          "422": {
            "description": "Validation failed"
          },
          "429": {
            "description": "Login limit"
          },
          "503": {
            "description": "Runtime or persistence unavailable"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/v1/config/plan": {
      "post": {
        "operationId": "post_config_plan",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "draft_revision": {
                    "minimum": 1,
                    "type": "integer"
                  }
                },
                "required": [
                  "draft_revision"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Operation completed"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "409": {
            "description": "Stale draft/plan"
          },
          "422": {
            "description": "Validation failed"
          },
          "429": {
            "description": "Login limit"
          },
          "503": {
            "description": "Runtime or persistence unavailable"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/v1/config/validate": {
      "post": {
        "operationId": "post_config_validate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "draft_revision": {
                    "minimum": 1,
                    "type": "integer"
                  }
                },
                "required": [
                  "draft_revision"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Operation completed"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "409": {
            "description": "Stale draft/plan"
          },
          "422": {
            "description": "Validation failed"
          },
          "429": {
            "description": "Login limit"
          },
          "503": {
            "description": "Runtime or persistence unavailable"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/v1/devices": {
      "get": {
        "operationId": "get_devices",
        "responses": {
          "200": {
            "description": "Redacted resource"
          },
          "401": {
            "description": "Authentication required"
          },
          "503": {
            "description": "Not ready or unavailable"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/v1/diagnostics": {
      "get": {
        "operationId": "get_diagnostics",
        "responses": {
          "200": {
            "description": "Redacted resource"
          },
          "401": {
            "description": "Authentication required"
          },
          "503": {
            "description": "Not ready or unavailable"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/v1/diagnostics/bundle": {
      "get": {
        "operationId": "get_diagnostics_bundle",
        "responses": {
          "200": {
            "content": {
              "application/gzip": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Redacted fixed-name diagnostics archive"
          },
          "401": {
            "description": "Authentication required"
          },
          "503": {
            "description": "Not ready or unavailable"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/v1/dns": {
      "get": {
        "operationId": "get_dns",
        "responses": {
          "200": {
            "description": "Redacted resource"
          },
          "401": {
            "description": "Authentication required"
          },
          "503": {
            "description": "Not ready or unavailable"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/v1/groups": {
      "get": {
        "operationId": "get_groups",
        "responses": {
          "200": {
            "description": "Redacted resource"
          },
          "401": {
            "description": "Authentication required"
          },
          "503": {
            "description": "Not ready or unavailable"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
    "/api/v1/logs": {
      "get": {
        "operationId": "get_logs",
        "responses": {
          "200": {
            "description": "Redacted resource"
          },
          "401": {
            "description": "Authentication required"
          },
          "503": {
            "description": "Not ready or unavailable"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/v1/openapi.json": {
      "get": {
        "operationId": "get_openapi_json",
        "responses": {
          "200": {
            "description": "Redacted resource"
          },
          "401": {
            "description": "Authentication required"
          },
          "503": {
            "description": "Not ready or unavailable"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/v1/proxies": {
      "get": {
        "operationId": "get_proxies",
        "responses": {
          "200": {
            "description": "Redacted resource"
          },
          "401": {
            "description": "Authentication required"
          },
          "503": {
            "description": "Not ready or unavailable"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
    "/api/v1/rules": {
      "get": {
        "operationId": "get_rules",
        "responses": {
          "200": {
            "description": "Redacted resource"
          },
          "401": {
            "description": "Authentication required"
          },
          "503": {
            "description": "Not ready or unavailable"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
    "/api/v1/subscriptions/inspect": {
      "post": {
        "operationId": "post_subscriptions_inspect",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "limit": {
                    "maximum": 128,
                    "minimum": 1,
                    "type": "integer"
                  },
                  "offset": {
                    "minimum": 0,
                    "type": "integer"
                  }
                },
                "required": [
                  "id",
                  "offset",
                  "limit"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Operation completed"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "409": {
            "description": "Stale draft/plan"
          },
          "422": {
            "description": "Validation failed"
          },
          "429": {
            "description": "Login limit"
          },
          "503": {
            "description": "Runtime or persistence unavailable"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/v1/subscriptions/refresh": {
      "post": {
        "operationId": "post_subscriptions_refresh",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "id": {
                    "pattern": "^[a-zA-Z0-9_-]{1,64}$",
                    "type": "string"
                  }
                },
                "required": [
                  "id"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Operation completed"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "409": {
            "description": "Stale draft/plan"
          },
          "422": {
            "description": "Validation failed"
          },
          "429": {
            "description": "Login limit"
          },
          "503": {
            "description": "Runtime or persistence unavailable"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/v1/system": {
      "get": {
        "operationId": "get_system",
        "responses": {
          "200": {
            "description": "Redacted resource"
          },
          "401": {
            "description": "Authentication required"
          },
          "503": {
            "description": "Not ready or unavailable"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/v1/system/recover": {
      "post": {
        "operationId": "post_system_recover",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Operation completed"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Authentication required"
          },
          "409": {
            "description": "Stale draft/plan"
          },
          "422": {
            "description": "Validation failed"
          },
          "429": {
            "description": "Login limit"
          },
          "503": {
            "description": "Runtime or persistence unavailable"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    }
  }
}